+++ src/test/java/com/sparta/ticketauction/domain/user/service/UserServiceImplTest.java
@@ -25,6 +25,7 @@ import com.sparta.ticketauction.domain.user.request.UserCreateRequest;
 import com.sparta.ticketauction.domain.user.request.UserPasswordUpdateRequest;
 import com.sparta.ticketauction.domain.user.request.UserUpdateRequest;
 import com.sparta.ticketauction.domain.user.response.UserResponse;
+import com.sparta.ticketauction.domain.user.service.impl.UserServiceImpl;
 import com.sparta.ticketauction.domain.user.util.UserUtil;
 import com.sparta.ticketauction.global.exception.ApiException;
 import com.sparta.ticketauction.global.util.LettuceUtils;
@@ -428,4 +429,39 @@ class UserServiceImplTest {
 				.hasMessage(ALREADY_USED_PASSWORD.getMessage());
 		}
 	}
+
+	@Nested
+	class 회원_탈퇴_테스트 {
+
+		@Test
+		void 성공() {
+			// Given
+			User user = TEST_USER;
+
+			given(userRepository.findByIdAndIsDeletedIsFalse(any())).willReturn(Optional.ofNullable(user));
+
+			// When
+			sut.deleteUser(user, 1L);
+
+			// Then
+			verify(userRepository).findByIdAndIsDeletedIsFalse(any());
+		}
+
+		@Test
+		void 권한_없음으로_실패() {
+			// Given
+			User user = TEST_USER;
+
+			// When
+			ApiException exception = assertThrows(
+				ApiException.class,
+				() -> sut.deleteUser(user, 2L)
+			);
+
+			// Then
+			assertThat(exception)
+				.hasMessage(ACCESS_DENIED.getMessage());
+		}
+	}
+
 }
